+++ packages/server/fileStorage/LocalFileStoreManager.ts
@@ -15,7 +15,11 @@ export default class LocalFileStoreManager extends FileStoreManager {
     }
   }
 
-  protected async putFile(file: ArrayBufferLike, fullPath: string) {
+  protected async putFile(
+    file: ArrayBufferLike,
+    fullPath: string,
+    _options?: {contentDisposition?: string}
+  ) {
     const fsAbsLocation = path.join(process.cwd(), fullPath)
     await fs.promises.mkdir(path.dirname(fsAbsLocation), {recursive: true})
     await fs.promises.writeFile(fsAbsLocation, Buffer.from(file) as any)